k8s: Refactor CNP status update logic into pkg/k8s
Refactor the CNP status code into the k8s package. This is required to
introduce exponential backoff and cluster size awareness. It is difficult to
make the existing K8sSuite unit tests aware of Daemon so by moving the status
update code into k8s, the dependency on the Daemon structure can be avoided.
